+++ source/renderer/app/components/wallet/transactions/TransactionTypeIcon.js
@@ -2,23 +2,75 @@
 import React, { Component } from 'react';
 import classNames from 'classnames';
 import SVGInline from 'react-svg-inline';
+import { LoadingSpinner } from 'react-polymorph/lib/components/LoadingSpinner';
+import { LoadingSpinnerSkin } from 'react-polymorph/lib/skins/simple/LoadingSpinnerSkin';
+import styles from './TransactionTypeIcon.scss';
+import spinnerOverrides from './SpinnerOverrides.scss';
 import expendIcon from '../../../assets/images/wallet-nav/send-ic.inline.svg';
 import incomeIcon from '../../../assets/images/wallet-nav/receive-ic.inline.svg';
 import exchangeIcon from '../../../assets/images/exchange-ic.inline.svg';
-import styles from './TransactionTypeIcon.scss';
-import { TransactionTypes } from '../../../domains/WalletTransaction';
+import pendingIcon from '../../../assets/images/wallet-nav/pending.inline.svg';
+import {
+  TransactionTypes,
+  TransactionStates,
+} from '../../../domains/WalletTransaction';
 
 type Props = {
+  exceedsPendingTimeLimit: boolean,
   iconType: string,
 };
 
 export default class TransactionTypeIcon extends Component<Props> {
+  applyIconStyles = (iconType: string): string => {
+    if (iconType !== TransactionStates.PENDING) {
+      return iconType;
+    }
+
+    if (!this.props.exceedsPendingTimeLimit) {
+      return `${iconType}_regular`;
+    }
+    return `${iconType}_warning`;
+  };
+
+  renderPendingIcon = () => {
+    if (!this.props.exceedsPendingTimeLimit) {
+      return this.renderPendingRegularIcon();
+    }
+    return this.renderPendingWarningIcon();
+  };
+
+  renderPendingRegularIcon = () => (
+    <div className={styles.pendingTxnIconWrapper}>
+      <LoadingSpinner
+        skin={LoadingSpinnerSkin}
+        themeOverrides={spinnerOverrides}
+      />
+    </div>
+  );
+
+  renderPendingWarningIcon = () => (
+    <div className={styles.pendingTxnIconWrapper}>
+      <LoadingSpinner
+        skin={LoadingSpinnerSkin}
+        themeOverrides={spinnerOverrides}
+      />
+      <SVGInline svg={pendingIcon} className={styles.pendingTxnIcon} />
+    </div>
+  );
+
+  renderIcon = (icon: string) => {
+    if (this.props.iconType === TransactionStates.PENDING) {
+      return this.renderPendingIcon();
+    }
+    return <SVGInline svg={icon} className={styles.transactionTypeIcon} />;
+  };
+
   render() {
     const { iconType } = this.props;
 
     const transactionTypeIconClasses = classNames([
       styles.transactionTypeIconWrapper,
-      styles[iconType],
+      styles[this.applyIconStyles(iconType)],
     ]);
 
     let icon;
@@ -38,9 +90,7 @@ export default class TransactionTypeIcon extends Component<Props> {
     }
 
     return (
-      <div className={transactionTypeIconClasses}>
-        <SVGInline svg={icon} className={styles.transactionTypeIcon} />
-      </div>
+      <div className={transactionTypeIconClasses}>{this.renderIcon(icon)}</div>
     );
   }
 }