+++ source/renderer/app/components/wallet/transactions/WalletTransactionsList.js
@@ -44,6 +44,7 @@ const messages = defineMessages({
 });
 
 type Props = {
+  deletePendingTransaction: Function,
   formattedWalletAmount: Function,
   hasMoreToLoad: boolean,
   isLoadingTransactions: boolean,
@@ -55,6 +56,7 @@ type Props = {
   showMoreTransactionsButton?: boolean,
   transactions: Array<WalletTransaction>,
   walletId: string,
+  isDeletingTransaction: boolean,
 };
 
 const DATE_FORMAT = 'YYYY-MM-DD';
@@ -178,10 +180,13 @@ export default class WalletTransactionsList extends Component<Props> {
 
   renderTransaction = (data: TransactionInfo): Node => {
     const {
+      deletePendingTransaction,
       formattedWalletAmount,
       isRestoreActive,
       network,
       onOpenExternalLink,
+      walletId,
+      isDeletingTransaction,
     } = this.props;
     const { isFirstInGroup, isLastInGroup, tx } = data;
     const txClasses = classnames([
@@ -193,6 +198,7 @@ export default class WalletTransactionsList extends Component<Props> {
       <div id={`tx-${tx.id}`} className={txClasses}>
         <Transaction
           data={tx}
+          deletePendingTransaction={deletePendingTransaction}
           formattedWalletAmount={formattedWalletAmount}
           isExpanded={this.isTxExpanded(tx)}
           isLastInList={isLastInGroup}
@@ -201,6 +207,8 @@ export default class WalletTransactionsList extends Component<Props> {
           onDetailsToggled={() => this.toggleTransactionExpandedState(tx)}
           onOpenExternalLink={onOpenExternalLink}
           state={tx.state}
+          walletId={walletId}
+          isDeletingTransaction={isDeletingTransaction}
         />
       </div>
     );