+++ source/renderer/app/containers/wallet/WalletSummaryPage.js
@@ -53,6 +53,8 @@ export default class WalletSummaryPage extends Component<Props> {
       recent,
       recentTransactionsRequest,
       unconfirmedAmount,
+      deletePendingTransaction,
+      deleteTransactionRequest,
     } = transactions;
     const wallet = wallets.active;
     // Guard against potential null values
@@ -77,7 +79,9 @@ export default class WalletSummaryPage extends Component<Props> {
           transactions={take(recent, MAX_TRANSACTIONS_ON_SUMMARY_PAGE)}
           isLoadingTransactions={recentTransactionsRequest.isExecutingFirstTime}
           hasMoreToLoad={false}
+          deletePendingTransaction={deletePendingTransaction}
           walletId={wallet.id}
+          isDeletingTransaction={deleteTransactionRequest.isExecuting}
           isRestoreActive={isRestoreActive}
           formattedWalletAmount={formattedWalletAmount}
           showMoreTransactionsButton={