+++ storybook/stories/WalletScreens.stories.js
@@ -171,10 +171,12 @@ storiesOf('WalletScreens', module)
           )
         ),
       ]}
+      deletePendingTransaction={() => {}}
       isLoadingTransactions={boolean('isLoadingTransactions', false)}
       isRestoreActive={boolean('isRestoreActive', false)}
       hasMoreToLoad={false}
       walletId="test-wallet"
+      isDeletingTransaction={false}
       formattedWalletAmount={formattedWalletAmount}
       totalAvailable={
         number('Transactions Sent', 1) + number('Transactions Received', 1)